refactor(semesters): moved createSemesterRequestBody to type
