+++ packages/shared/src/types/index.ts
@@ -3,3 +3,4 @@ export * from "./enums"
 export * from "./game-session"
 export * from "./member"
 export * from "./auth"
+export * from "./semester"
